requirements.txt
joblib
scikit-learn
numpy
SpeechRecognition
gTTS
pydub
openai-whisper
requests
pyaudio
